backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
playwright
python-dotenv